feat(host): E2 接线 — host-agent 数据流监控 + streamer 推流状态发布
- host-agent: --token <文件>（Monitor 角色令牌）→ FlowMonitor attach + 5s 循环
  日志 FlowSnapshot（topic fps/bps/stalled + stream 推流状态）；无 token 跳过
  （拓扑监控不受影响，B 阶段 oxfile 签发后启用）
- host-streamer: log_stats additive 发布 StreamerStats JSON 到 stats/stream-<id>
  （FrameMeta::FORMAT_JSON + 单调 ts_mono + seq；保留原日志）
- ACL 定稿: Pusher 保持单方向（host.rs build_acl 契约, 显式 --topic 支持）;
  stats/* 发布权仅 Recorder（streamer 令牌缺省 Recorder, C2 遗留）
- 回归: host 104/104 + link 全绿 + token_cli 9/9 + pixi run check 0 error

+++ crates/mediaservo-host/src/monitor/flow.rs
@@ -53,8 +53,8 @@ pub struct StreamFlow {
     pub id: String,
     /// 最近一次 stats 的 bytes_sent（webrtc OutboundRtp，累计）。
     pub bytes_sent: u64,
-    /// 最近一次 stats 的 frames_encoded（累计）。
-    pub frames_encoded: u64,
+    /// 最近一次 stats 的 frames_encoded（libwebrtc u32，累计）。
+    pub frames_encoded: u32,
     pub frame_width: u32,
     pub frame_height: u32,
     /// 最近 stats 是否在新鲜窗口内。
@@ -72,7 +72,8 @@ pub struct FlowSnapshot {
 #[derive(Debug, Clone, Copy, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
 pub struct StreamerStats {
     pub bytes_sent: u64,
-    pub frames_encoded: u64,
+    /// 与 libwebrtc RTCOutboundRtpStreamStats 对齐（u32）。
+    pub frames_encoded: u32,
     pub frame_width: u32,
     pub frame_height: u32,
 }
